disable code signing and fix native advert
- We don't have the 2010 certificate signed browsercontrol dlls so the
code signing routines are failing with the new certificates, so we
needed to update the check for certificate chain length. We also had to
remove the hardcoded checks for serialNumber and public key.After doing
that the old certificates were then failing because of the chain length.

- When updating the package name for the native advert class the JNI
methods were failing to be called since we couldn't resolve the names.
Updating the package name fixed this

Signed-off-by: Open592 Developer <uss@open592.com>