ContextualEntry: Initial name for this extention of LinkedEntry
Not sure if this is the best fit, but makes the most sense to me as I'm
not sure I can think of another reason to extend the entry with the same
fields.

Thinking of it as basically a object oriented version of this:

struct Entry {
  long hashKey;
  Gender gender;
  Entry *next;
  Entry *previous;
  Entry *nextMale;
  Entry *previousMale;
}

There is probably another class somewhere in here which adds yet another
bundle of pointers. Will have to reevaluate the naming once I think
through that one.

Signed-off-by: Open592 Developer <uss@open592.com>