Node: Simplify the naming of some of this.
I've been going back and forth, and this is the best I've come up with
to this point. Leans a bit on what OpenRS2 has used, but after going
through each of these and seeing their usages, I agree with their
approach.

As always, after going through the code more, I may come up with a
different scheme. But this seems ok for now.

Signed-off-by: Open592 Developer <uss@open592.com>