SignLink: Start adding back message handling from loader.
I extracted the SignLink class which existed in the client instead of
using the one which existed in the loader. The client has much less
capabilities due to it operating outside a signed context.

There are still a couple more messsages to go and further refactoring,
but this is a good start.

Signed-off-by: Open592 Developer <uss@open592.com>