Packet: Fix bug in g4()
This was a really complicated bug to track down, and was presenting
itself as a CRC mismatch which made me spend way too much time
investigating file server logic. But in the end, I tracked it down to a
client issue since the files served by the file server had the correct
generated CRC, and the ukeys on the server were generating the correct
CRC as well.

Was also really weird because we use g4() to grab archive version and
that was also correctly being read... but this bug was only detected
when reading an integer with the most significant bits filled (which the
version number did not)

This was obviously not a client bug, but was instead the result of some
refactoring I did a very long time ago when renaming the packet methods.
Dumb me...

Once this bug was fixed (and all the other file server bugs which I
fixed along the way) we make it all the way to the world server packet!
And got a glimpse of the first RS art! Amazing!

Signed-off-by: Open592 Developer <uss@open592.com>